Remove the last -Wconversion warnings from all Mesh_3
I have only compiled tests and examples. Maybe the Mesh_3 plugin of the 3D
demo has other warnings.

+++ Mesh_3/examples/Mesh_3/mesh_3D_gray_image_multiple_values.cpp
@@ -55,7 +55,7 @@ int main(int argc, char*argv[])
     iso_values.insert(2.9f);
   } else {
     for(int i = 2; i < argc; ++i) {
-      iso_values.insert(std::atof(argv[i]));
+      iso_values.insert(static_cast<float>(std::atof(argv[i])));
     }
   }
